# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Book:
    properties:
      author:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      isbn:
        type: string
      language:
        type: string
      pages:
        type: integer
      price:
        type: integer
      published_at:
        type: string
      stock:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  domain.Error:
    properties:
      code:
        type: integer
      errors:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  domain.Success:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  domain.TransactionDetail:
    properties:
      book:
        $ref: '#/definitions/domain.Book'
      book_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      quantity:
        type: integer
      sub_total:
        type: integer
      transaction_id:
        type: integer
      updatedAt:
        type: string
    required:
    - book_id
    - quantity
    - sub_total
    - transaction_id
    type: object
  domain.TransactionDetailStoreRequest:
    properties:
      book_id:
        type: integer
      quantity:
        type: integer
    required:
    - book_id
    - quantity
    type: object
  domain.TransactionStoreRequest:
    properties:
      customer_id:
        type: integer
      transaction_details:
        items:
          $ref: '#/definitions/domain.TransactionDetailStoreRequest'
        type: array
      user_id:
        type: integer
    required:
    - customer_id
    - transaction_details
    - user_id
    type: object
  domain.TransactionUpdateRequest:
    properties:
      customer_id:
        type: integer
      transaction_details:
        items:
          $ref: '#/definitions/domain.TransactionDetail'
        type: array
      user_id:
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
host: localhost:3000
info:
  contact: {}
  title: Book Store API Documentation
  version: "1.0"
paths:
...
  /transactions:
    get:
      consumes:
      - application/json
      description: Get list of transactions
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Size of page (default 10)
        in: query
        name: size
        type: integer
      - description: Customer Id
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of transactions
          schema:
            items:
              $ref: '#/definitions/domain.Success'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get list of transaction
      tags:
      - transactions
    post:
      consumes:
      - application/json
      description: Store transaction
      parameters:
      - description: transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/domain.TransactionStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: transaction detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Store transaction
      tags:
      - transactions
  /transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Delete transaction
      parameters:
      - description: transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success delete transaction
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Delete transaction
      tags:
      - transactions
    get:
      consumes:
      - application/json
      description: Get transaction by id
      parameters:
      - description: transaction ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: transaction detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get transaction by id
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Update transaction
      parameters:
      - description: transaction ID
        in: path
        name: id
        required: true
        type: integer
      - description: transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/domain.TransactionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: transaction detail
          schema:
            $ref: '#/definitions/domain.Success'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      security:
      - Bearer: []
      summary: Update transaction
      tags:
      - transactions
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"